refactor: reorganize infrastructure modules
